Add comments to all functions. Add check on params to deleteCharacter. Add check for user in getCharacterDetails.

+++ src/controllers/characters-controller.js
@@ -1,5 +1,12 @@
 const knex = require("knex")(require("../../knexfile.js"));
 
+/**
+ * getAllUsersCharacters is a function that fufills a GET request. All of a user's characters will be returned in an array. The data is reduced to the characters id, name, and background name.
+ * 
+ * @param {Object}      req 
+ * @param {Object}      res 
+ * 
+ */
 const getAllUsersCharacters = async (req, res) => {
     try {
         // TODO check if user_id exists in database?
@@ -21,85 +28,111 @@ const getAllUsersCharacters = async (req, res) => {
     }
 }
 
+/**
+ * getCharacterDetails is a function that fufills a GET request. All of a character's data (in the characters table and relational tables using the charact_id as a foreign key) will be returned in the response.
+ * 
+ * @param {Object}      req 
+ * @param {Object}      res 
+ * 
+ */
 const getCharacterDetails = async (req, res) => {
     try{
-        // TODO check if user_id exists in database?
+        // Check if user_id+username exists in database
         const userId = req.decode.user_id;
-        // Check to see if the ID of the inventory is valid
-        const characterId = req.params.id;
-        if (!characterId || isNaN(characterId)) {
-            // Return the response of an invalid ID
-             return res.status(400).send(`Invalid character id: ${characterId}`);
-        }
+        const userName = req.decode.username;
+        const referencedUser = await knex("users").where({
+            user_id: userId,
+            username: userName 
+        });
 
-        //Get a list of the characters belonging to this user. Only return the information on the character's name, id, and their background's name 
-        const characterData = await knex('characters')
-        .join("backgrounds", "backgrounds.background_id", "characters.background_id")
-        .select(
-            "characters.character_id",
-            "characters.name AS character_name",
-            "characters.gender",
-            "characters.age",
-            "characters.birthplace",
-            "characters.residence",
-            "characters.background_id",
-            "backgrounds.name AS background_name",
-            "characters.special_people",
-            "characters.favoured_possession",
-            "characters.mania",
-            "characters.traits",
-            "characters.injuries",
-            "characters.feats",
-            "characters.spells",
-            "characters.notes"            
-        )
-        .where({user_id: userId, 'characters.character_id': characterId})
-        .first();
+        if (referencedUser.length === 0) {
+            return res.status(400).json({
+              error: `Request cannot be completed as the user ${userName} does not exist.`,
+            });
 
-        if(!characterData){
-            // Return the results of the query
-            res.status(404).json({error: `No character with id ${characterId} belongs to user with id ${userId}`});
-        }else{
-            // Get the character's stats as an object
-            const statsData = await knex('stats')
-            .select("*")          
-            .where({character_id: characterId})
+        } else {
+            // Check to see if the ID of the character is valid
+            const characterId = req.params.id;
+            if (!characterId || isNaN(characterId)) {
+                // Return the response of an invalid ID
+                return res.status(400).send(`Invalid character id: ${characterId}`);
+            }
+
+            //Get a list of the characters belonging to this user. Only return the information on the character's name, id, and their background's name 
+            const characterData = await knex('characters')
+            .join("backgrounds", "backgrounds.background_id", "characters.background_id")
+            .select(
+                "characters.character_id",
+                "characters.name AS character_name",
+                "characters.gender",
+                "characters.age",
+                "characters.birthplace",
+                "characters.residence",
+                "characters.background_id",
+                "backgrounds.name AS background_name",
+                "characters.special_people",
+                "characters.favoured_possession",
+                "characters.mania",
+                "characters.traits",
+                "characters.injuries",
+                "characters.feats",
+                "characters.spells",
+                "characters.notes"            
+            )
+            .where({user_id: userId, 'characters.character_id': characterId})
             .first();
 
-            // Remove the unnecessary keys from the object
-            delete statsData.character_id;
-            delete statsData.created_at;
-            delete statsData.updated_at;
+            if(!characterData){
+                // Return the results of the query
+                res.status(404).json({error: `No character with id ${characterId} belongs to user with id ${userId}`});
+            }else{
+                // Get the character's stats as an object
+                const statsData = await knex('stats')
+                .select("*")          
+                .where({character_id: characterId})
+                .first();
 
-            // Get an array of all the character's skill points
-            const skillsData = await knex('skill_points')
-            .select(
-                "skill_points.skill_id",
-                "skill_points.points",
-            )          
-            .where({character_id: characterId});
+                // Remove the unnecessary keys from the object
+                delete statsData.character_id;
+                delete statsData.created_at;
+                delete statsData.updated_at;
 
-             // Get an array of all the character's equipment
-             const equipmentData = await knex('equipment')
-             .select(
-                 "equipment.equipment_id",
-                 "equipment.name",
-             )          
-             .where({character_id: characterId});
-            
-            // Add them onto the characterData object
-            const characterSheet = {...characterData, stats: statsData, skills: skillsData,  equipment: equipmentData};
-            
-            // Return the results of the query
-            res.status(200).json(characterSheet);
+                // Get an array of all the character's skill points
+                const skillsData = await knex('skill_points')
+                .select(
+                    "skill_points.skill_id",
+                    "skill_points.points",
+                )          
+                .where({character_id: characterId});
+
+                // Get an array of all the character's equipment
+                const equipmentData = await knex('equipment')
+                .select(
+                    "equipment.equipment_id",
+                    "equipment.name",
+                )          
+                .where({character_id: characterId});
+                
+                // Add them onto the characterData object
+                const characterSheet = {...characterData, stats: statsData, skills: skillsData,  equipment: equipmentData};
+                
+                // Return the results of the query
+                res.status(200).json(characterSheet);
+            }
         }
-
 
     }catch(error){
         res.status(500).send(`Error retrieving character's details: ${error}`);
     }
 }
 
+/**
+ * addCharacter is a function that fufills a POST request. A character will be created in the characters table, as well as populating the data related to the character in other related tables.
+ * 
+ * @param {Object}      req 
+ * @param {Object}      res 
+ * 
+ */
 const addCharacter = async (req, res) => {
     try{
         // Check if user_id+username exists in database
@@ -217,6 +250,13 @@ const addCharacter = async (req, res) => {
     }
 }
 
+/**
+ * editCharacter is a function that fufills a PUT request. A character will be updated in the characters table, as well as all relevent data in other tables that use that character's character_id as a foreign key.
+ * 
+ * @param {Object}      req 
+ * @param {Object}      res 
+ * 
+ */
 const editCharacter = async (req, res) => {
     try{
         // Check if user_id+username exists in database
@@ -232,11 +272,14 @@ const editCharacter = async (req, res) => {
             });
 
         }else {
+            // Check that the id is a valid id
             const characterId = req.params.id;
             if (!characterId || isNaN(characterId)) {
                 // Return the response of an invalid ID
                  return res.status(400).send(`Invalid character id: ${characterId}`);
             }
+
+            //Check that a character with this id exists for this user 
             const existingCharacter = await knex('characters')
             .where({character_id: characterId, user_id: userId})
             .first();
@@ -363,6 +406,13 @@ const editCharacter = async (req, res) => {
     }
 }
 
+/**
+ * deleteCharacter is a function that fufills a DELETE request. A character will be deleted from the characters table, as well as all references in other tables that use that character's character_id as a foreign key.
+ * 
+ * @param {Object}      req 
+ * @param {Object}      res 
+ * 
+ */
 const deleteCharacter = async (req, res) => {
     try {
         // Check if user_id+username exists in database
@@ -378,13 +428,20 @@ const deleteCharacter = async (req, res) => {
             });
 
         }else {
+            // Check that the id is a valid id
+            const characterId = req.params.id;
+            if (!characterId || isNaN(characterId)) {
+                // Return the response of an invalid ID
+                 return res.status(400).send(`Invalid character id: ${characterId}`);
+            }
+
             const rowsDeleted = await knex("characters")
-            .where({ character_id: req.params.id, user_id: userId })
+            .where({ character_id: characterId, user_id: userId })
             .delete();
 
             if (rowsDeleted === 0) {
                 // Nothing was deleted as the character wasn't found associated with that user
-                res.status(404).json({ error: `Character ${req.params.id} not found under user ${userName} to delete.` });
+                res.status(404).json({ error: `Character ${characterId} not found under user ${userName} to delete.` });
             }else{
                 // No Content response for success
                 res.sendStatus(204);
